# Move row scanning int the data chunk interface